Restructurize tests
