+++ src/test/java/pl/put/poznan/buildingInfo/logic/StructureRepositoryTest.java
@@ -1,0 +1,12 @@
+package pl.put.poznan.buildingInfo.logic;
+
+import org.junit.Before;
+
+import static org.junit.Assert.*;
+
+public class StructureRepositoryTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+}
